Cleanup for MethodRefLinker Metadata language specific Dereference
