+++ joern-cli/frontends/x2cpg/src/main/scala/io/joern/x2cpg/passes/base/FileCreationPass.scala
@@ -6,14 +6,15 @@ import io.shiftleft.codepropertygraph.generated.{EdgeTypes, NodeTypes, PropertyN
 import io.shiftleft.passes.CpgPass
 import io.shiftleft.semanticcpg.language._
 import io.shiftleft.semanticcpg.language.types.structure.FileTraversal
-import io.joern.x2cpg.passes.callgraph.MethodRefLinker
+import io.joern.x2cpg.utils.LinkingUtil
 
 import scala.collection.mutable
 
 /** For all nodes with FILENAME fields, create corresponding FILE nodes and connect node with FILE node via outgoing
   * SOURCE_FILE edges.
   */
-class FileCreationPass(cpg: Cpg) extends CpgPass(cpg) {
+class FileCreationPass(cpg: Cpg) extends CpgPass(cpg) with LinkingUtil {
+
   override def run(dstGraph: DiffGraphBuilder): Unit = {
     val originalFileNameToNode = mutable.Map.empty[String, StoredNode]
     val newFileNameToNode      = mutable.Map.empty[String, NewFile]
@@ -37,10 +38,8 @@ class FileCreationPass(cpg: Cpg) extends CpgPass(cpg) {
       }
     }
 
-    // Create SOURCE_FILE edges from nodes of various types
-    // to FILE
-
-    MethodRefLinker.linkToSingle(
+    // Create SOURCE_FILE edges from nodes of various types to FILE
+    linkToSingle(
       cpg,
       srcLabels = List(NodeTypes.NAMESPACE_BLOCK, NodeTypes.TYPE_DECL, NodeTypes.METHOD, NodeTypes.COMMENT),
       dstNodeLabel = NodeTypes.FILE,
@@ -53,4 +52,5 @@ class FileCreationPass(cpg: Cpg) extends CpgPass(cpg) {
       Some(createFileIfDoesNotExist)
     )
   }
+
 }
