+++ joern-cli/frontends/x2cpg/src/main/scala/io/joern/x2cpg/passes/base/NamespaceCreator.scala
@@ -23,4 +23,5 @@ class NamespaceCreator(cpg: Cpg) extends CpgPass(cpg) {
         blocks.foreach(block => dstGraph.addEdge(block, namespace, EdgeTypes.REF))
       }
   }
+
 }
